Add all the review comments from PR #40
