--- pipelines/clickstream_analytics_java/app/build.gradle.kts
@@ -1,41 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Java application project to get you started.
- * For more details on building Java & JVM projects, please refer to https://docs.gradle.org/8.2.1/userguide/building_java_projects.html in the Gradle documentation.
- */
-
-plugins {
-    application
-}
-
-repositories {
-    mavenCentral()
-    maven {
-            url = uri("https://packages.confluent.io/maven/")
-        }
-}
-
-dependencies {
-    testImplementation("junit:junit:4.13.2")
-    implementation("com.google.guava:guava:31.1-jre")
-    implementation("com.google.cloud.dataflow:google-cloud-dataflow-java-sdk-all:2.5.0")
-    implementation("org.apache.beam:beam-sdks-java-io-google-cloud-platform:2.50.0")
-    implementation("org.apache.beam:beam-sdks-java-core:2.50.0")
-    implementation("org.apache.beam:beam-runners-google-cloud-dataflow-java:2.50.0")
-    annotationProcessor("com.google.auto.value:auto-value-annotations:1.10.4")
-    implementation("com.google.auto.value:auto-value-annotations:1.10.4")
-    implementation("org.slf4j:slf4j-nop:2.0.13")
-    implementation("io.confluent:kafka-avro-serializer:5.3.2")
-    implementation("io.confluent:kafka-schema-registry-client:5.3.2")
-}
-
-java {
-    toolchain {
-        languageVersion.set(JavaLanguageVersion.of(17))
-    }
-}
-
-application {
-    mainClass.set("clickstream_analytics_java.ClickstreamPubSubToBq")
-}
