--- pipelines/clickstream_analytics_java/app/src/main/java/clickstream_analytics_java/BigTableEnrichment.java
@@ -1,8 +1,0 @@
-package clickstream_analytics_java;
-
-
-public class BigTableEnrichment{
-
-    /*** TODO ***/
-
-}
