+++ pipelines/clickstream_analytics_java/src/main/resources/streaming_source_deadletter_table_schema.json
